system-requirements.txt
python3-picamera2
libcamera-apps
python3-flask
python3-numpy
python3-pil
